# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Clashtools Backend API
  description: API Dokumentation für das Clashtools Backend
  contact:
    name: Thomas Standegger
    url: https://htl-kaindorf.at
    email: statha21@htl-kaindorf.at
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
  /api/auth/logout:
    post:
      tags:
        - Authentication Controller
      summary: Logs out the user
      description: Logs out the currently authenticated user and invalidates their session.
      operationId: logout
      parameters:
        - name: userDetails
          in: header
          description: Currently authenticated user details
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully logged out
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized, user not authenticated
        '404':
          description: User not found
  /api/auth/refresh:
    post:
      tags:
        - Authentication Controller
      summary: Refresh JWT Token
      description: Refreshes the JWT token using a valid refresh token.
      operationId: refreshToken
      requestBody:
        description: Refresh token request containing the refresh token
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequestDTO'
        required: true
      responses:
        '200':
          description: Successfully refreshed token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDTO'
        '400':
          description: Invalid or expired refresh token
  /api/auth/register:
    post:
      tags:
        - Authentication Controller
      summary: Register a new user
      description: Registers a new user with the provided registration details.
      operationId: register
      requestBody:
        description: Registration details for the new user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestDTO'
        required: true
      responses:
        '200':
          description: User successfully registered
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Invalid registration data or username already exists
  /api/auth/login:
    post:
      tags:
        - Authentication Controller
      summary: Login user
      description: Authenticates a user and returns a JWT token.
      operationId: login
      requestBody:
        description: Login credentials
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestDTO'
        required: true
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDTO'
        '401':
          description: Invalid credentials
        '400':
          description: Invalid request data
...
  /api/attributes/{id}:
    get:
      tags:
        - Attribute
      summary: Get attribute by ID
      description: Retrieves an attribute by its ID
      operationId: getAttributeById
      parameters:
        - name: id
          in: path
          description: ID of the attribute to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved the attribute
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeResponseDTO'
        '404':
          description: Attribute not found
    put:
      tags:
        - Attribute
      summary: Update an attribute
      description: Updates an existing attribute by its ID
      operationId: updateAttribute
      parameters:
        - name: id
          in: path
          description: ID of the attribute to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated attribute data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttributeRequestDTO'
        required: true
      responses:
        '200':
          description: Successfully updated the attribute
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttributeResponseDTO'
        '400':
          description: Invalid input data
        '404':
          description: Attribute not found
    delete:
      tags:
        - Attribute
      summary: Delete an attribute
      description: Deletes an attribute by its ID
      operationId: deleteAttribute
      parameters:
        - name: id
          in: path
          description: ID of the attribute to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successfully deleted the attribute
        '404':
          description: Attribute not found
...
components:
  schemas:
    AuthRequestDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponseDTO:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
      required:
        - token
        - refreshToken
    RefreshTokenRequestDTO:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    RegisterRequestDTO:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - username
        - email
        - password
...
    AttributeNameDTO:
      type: object
      description: DTO for Attribute Name information
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the attribute name
          example: 1
        name:
          type: string
          description: Name of the attribute
          example: damage_per_second
    AttributeTranslationDTO:
      type: object
      description: DTO for Attribute Translation information
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the attribute translation
          example: 1
        languageCode:
          type: string
          description: Language code for the translation
          example: en
        name:
          type: string
          description: Translated name of the attribute
          example: Damage per Second
      required:
        - languageCode
        - name
    AttributeRequestDTO:
      type: object
      description: Request DTO for creating or updating an Attribute
      properties:
        attributeNameId:
          type: integer
          format: int64
          description: ID of the attribute name
          example: 1
        translations:
          type: array
          description: List of attribute translations
          items:
            $ref: '#/components/schemas/AttributeTranslationDTO'
      required:
        - attributeNameId
    AttributeResponseDTO:
      type: object
      description: Response DTO for Attribute information
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the attribute
          example: 1
        attributeName:
          $ref: '#/components/schemas/AttributeNameDTO'
          description: The attribute name object
        translations:
          type: array
          description: List of attribute translations
          items:
            $ref: '#/components/schemas/AttributeTranslationDTO'
...